+++ build.gradle
@@ -3,6 +3,32 @@ plugins {
     id 'application'
 }
 
+
+
+
+
+
+plugins {
+    id 'com.github.johnrengelman.shadow' version '8.1.1'
+    id 'java'
+}
+
+dependencies {
+    shadow localGroovy()
+    shadow gradleApi()
+
+    implementation 'org.jdom:jdom2:2.0.6'
+    implementation 'org.ow2.asm:asm:6.0'
+    implementation 'org.ow2.asm:asm-commons:6.0'
+    implementation 'commons-io:commons-io:2.4'
+    implementation 'org.apache.ant:ant:1.9.4'
+    implementation 'org.codehaus.plexus:plexus-utils:2.0.6'
+}
+
+tasks.named('shadowJar', ShadowJar) {
+    enableRelocation true
+}
+
 group 'com.alebit.taikoweb'
 version '1.0.1'
