+++ src/main/kotlin/com/alebit/taikoweb/parser/GenreSimpleParser.kt
