+++ src/main/kotlin/com/alebit/taikoweb/struct/CsrfToken.kt
