+++ src/main/kotlin/com/alebit/taikoweb/struct/Genre.kt
