+++ src/main/kotlin/com/alebit/taikoweb/struct/Login.kt
