+++ src/main/kotlin/com/alebit/taikoweb/struct/LoginResponse.kt
