+++ src/main/kotlin/com/alebit/taikoweb/struct/TaikoWebSong.kt
